Decode immediates into i32
This simplifies our code and brings us closer to the text and tables of
the spec.

From https://riscv.org/wp-content/uploads/2017/05/riscv-spec-v2.2.pdf
page 11:

---

Except for the 5-bit immediates used in CSR instructions
(Section 2.8), immediates are always sign-extended, and are generally packed towards the leftmost
available bits in the instruction and have been allocated to reduce hardware complexity. In partic-
ular, the sign bit for all immediates is always in bit 31 of the instruction to speed sign-extension
circuitry.

---

See also the tables on page 12 which also talk about the 'immediates'
being full 32 bit conceptually.