feat: cross table lookup setup (#240)
_original PR: https://github.com/0xmozak/mozak-vm/pull/189_

This PR introduces the groundwork to support cross table lookups.

This PR does:
- introduces infra for cross table lookups (table instantiation,
columns, etc.)
- introduces generic tests for cross table lookups (simple manual CTL
checks)

**This PR does not**:
- actually use the cross table lookups for proof generation (we need to
prove a table with the lookup data first). Currently our CPU proof is
independent of lookup data, for example.

To address what this PR does not do I will be opening a separate PR for
that.