feat(lookups): implement prove and verify with lookups (#257)
This commit implements the proving and verifying of a table (mostly lifted from `starky`
crate) that includes lookup data, but doesn't use those yet. This commit still uses the prove and verify functions from `starky`'s public API to be able to prove/verify the CPU table. While the rangecheck table exists, it is not being used in this commit.