feat(lookups): prove/verify ADD with cpu<>rangecheck (#305)
This PR replaces usage of the default `prove` and `verify` from `starky`
crate, with our own `prove` and `verify` which includes cross table
lookups. We also add the usage of `RangeCheckStark` here, so we're now
proving and verifying both tables. Note that what we're proving is the ADD
instruction, and nothing else, since this serves as a POC of how rangechecks
cross table lookups can be done.

---------

Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>