refactor: remove unused rangecheck code (#333)
The rangecheck stark's `eval_ext_circuit` is never used in our use case
at this moment, so let's remove it to reduce diffs in future rangecheck
PRs.

The test `test_rangecheck_stark_circuit` is also removed since we do not
have the `eval_ext_circuit` implemented for this. You can see from #331
that there is duplicate implementation between `eval_packed_generic` and
`eval_ext_circuit` to satisfy this test.

Consequently `eval_lookups_circuit` is also unused, so we remove it as
well.

We can always add these back as needed in future.