refactor(rangecheck): generate rangecheck trace from cpu trace
All of our rangechecks are derived from the CPU trace, so we might as
well generate our rangecheck trace from the CPU trace. In future we can
add more traces to generate our rangecheck trace as necessary.

Note that this will also change our behaviour of our rangecheck trace
generation slightly - instead of indices we push to `Vec`s instead -
this is because our rangecheck trace may not have the same length as our
cpu trace, and in future we may have multiple rows rangechecking the
same CPU row.