Use a 'View' instead of manually calculating indices for columns (#382)
We adopt a technique from plonky2 to make our column access less
error-prone.

This PR only moves the constraints properly to the new system, and
mostly leaves trace generation as is; because this PR is already big
enough as it is. The next PR will move traces to the new system, too.