test: use native starky's prove and verify for most tests
Currently, we prove and verify a single `MozakStark` for all tests.
As we have added many more tests recently there's an obvious slowdown
in performance of our tests, because we are proving and verifying all
STARKs in all tests, even though most tests are 'unit tests' in the
sense that they are only concerned with proving and verifying a
single operation.

This PR renames the existing `simple_proof_test` function to
`prove_and_verify_mozak_stark`, and adds a new
`prove_and_verify_single_stark` function (self-explanatory names), and
switches most of tests to use the latter.

The obvious upside of this PR is that we get a significant improvement
in speed (nearly 10x for me; details below), but the downside is that in the
tests that use the new function, we are not currently proving and
verifying cross table lookups (which the original function did).

My argument for this PR would be that we can be more conscious about
which function to use, and only use `prove_and_verify_mozak_stark`
for ops that require lookups. This PR would also significant improve
the feedback loop for us (make changes -> compile -> test -> repeat).