refactor(tests): use `simple_test_code` everywhere; localize usage of `simple_test` (#430)
There are inconsistent usages of `simple_test_code` and `simple_test` in
our codebase. Since it seems like we're using `simple_test_code`
everywhere now, we can move `simple_test` to `vm.rs`. Note that we can't
simply remove `simple_test`, because we'd still like to read a raw
instruction for unimplemented opcodes (csrrs, fence, etc.), but this PR
makes it so we can move `simple_test` out of `test_utils.rs` into
`vm.rs` to localise it where it is used.

Note that we also make `simple_test` a non public function to make it
clear that `simple_test_code` is to be preferred.