feat(generation): support mem ops in cpu trace gen (#394)
works towards #378

## Constraints

The sumcheck for op2 = rs2 + imm stays the same, but this PR adds an
additional check for a possible overflow when it comes to memory
operations.

First, we introduce a new column, `op2_value_overflowing`. This
represents the sum `rs2 + imm` in **field element form**. This means
that the sum **can** overflow (in terms of u32). We need a new
constraint to check this overflow.

We have to check 2 cases - the overflowing case and the non-overflowing
case. Here's the full constraint:

```rust
    yield_constr.constraint(
        (lv.op2_value_overflowing - lv.op2_value)
            * (lv.op2_value_overflowing - lv.op2_value - wrap_at * lv.inst.ops.is_mem_op()),
    );
```

The non-overflowing case is the simpler case: in that scenario, we
simply check

```rust
    yield_constr.constraint(
        (lv.op2_value_overflowing - lv.op2_value)
            * ...
```

The overflowing case is slightly trickier:

```rust
    let wrap_at = lv.shifted(32);
    ...

    yield_constr.constraint(
        ...
            * (lv.op2_value_overflowing - lv.op2_value - wrap_at * lv.inst.ops.is_mem_op()),
    );
```

In this scenario we want to subtract by the full wrap amount, if the
current operation is a memory operation. In future PRs this should be
extended to include other memory instructions.

---------

Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>