fix(test): split rangecheck tests to check sum and range
Previously, our simple test was only checking for the sum constraint
and not the range constraint by altering a value to be larger than u32.
As a result it wasn't encapsulating the whole constraint system for range checks.

This PR refactors the tests by splitting it into two tests:

`test_rangecheck_stark_fails_sum_constraint` (original),
`test_rangecheck_stark_fails_range_constraint` (new).

The sum constraint test like before alters a value with limbs that are
within a u16 range, but we alter a limb directly this time to be more explicitly.

The range constraint test alters a limb to be larger than u16 (therefore
failing the range check), while still passing the sum constraint.