fix(generation): bitshift trace gen was excluding cpu bitshift ops (#578)
closes #577 

When doing `merge_join_by`, it is not enough to check for `is_left` - we
also require a check on `is_both`, otherwise the first match gets
skipped.

This commit also contains a cosmetic change, changing the name of
`dummy_or_executed` to `executed_or_dummy`. I'm assuming the dummy
refers to the 0..32 range, while executed refers to the actual executed
trace.