Range check to constrain sign bits (#593)
Implement the range check mentioned as TODO in `signed_comparison.rs` 

Notes
- jalr proptests are still WIP. There are just few handwritten examples.
- `prove_lbu` fails with MozakStark. The reason is that the constraints
for our memory table do not (yet) support reading from memory without
first initialising it with a write. We need to fix that in a subsequent
PR. @supragya is working on that as part of the wider
memory-initialisation-from-ELF effort. The idea is to initialise from
ELF, when the ELF mentions a particular address, and otherwise
initialise with 0.

---------

Co-authored-by: Matthias Goergens <matthias.goergens@gmail.com>
Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>