refactor: rename `vm` sub-crate to `runner` (#623)
This whole repository is called the Mozak-VM. But we also have a
sub-crate called `mozak-vm` in the directory `vm`. That's rather
confusing when talking about our system.

In this PR, we rename the directory from `vm` to `runner`, and the
sub-create from `mozak-vm` to `mozak-runner`.

I also considered other names, like `executor`, but `runner` is simpler
and easier to pronounce.