refactor: make signedness its own column, unify operations (#611)
This saves a few columns in the cpu and program ROM tables, and
implements the long-standing TODO of unify where we specify signedness
of operands.

We also save a few rangechecks by more precisely specifying when the
sign bits need to be constrained via ctl. That was possible before this
refactor, but would have been a bit more annoying to do and would have
duplicated logic.