feat(register): introduce `RegisterInitStark`
Works towards https://github.com/0xmozak/mozak-vm/issues/487.

Part of the effort towards treating our register file like memory.

We require an init STARK here in order to have an auxiliary table to check
certain constraints in the `RegisterStark` that we will eventually have.
For example, in certain cases, we have to constrain that certain values
do not change across rows, and we will use a sorted table with the init
row being the first row to segregate sections.