feat(reg): Introduce `diff_augmented_clk` column
Introduces a new column to the `RegisterStark` in order to prevent
exploits that swap `augmented_clk`s around in the trace.

This PR also changes the calculation of the `augmented_clk` - instead of
`augmented_clk = clk * 2`, we do  `augmented_clk = clk * 3` (so we have
space to differentiate between `rs1` and `rs2` reads.