fix(register): fix "rd must not change" constraint
Previously, all we did was constrain by filtering the local `is_read`.
This wasn't sufficient because the next filter can easily be `is_init`,
and the values might change (and that's valid).

We need to constrain by the negation of `is_write`, and on top of that
also add a constrain on the address, since it is perfectly valid for
a register to be read followed by an init row of the next register.