Reduce opt-level
Experiments on my machine show that the tests run nearly as fast with
opt-level 1 as with opt-level 3, but build times are a bit shorter (from
a `cargo clean` start.)

Please run some experiments on your machine, and contribute data.

For opt-level 1, 2 and 3 the tests take roughly 2s to run on my machine.

By contrast, with opt-level 0 the tests take 23s to run.