refactor: remove all `generic_const_exprs` usage (#725)
This PR gets rid of `generic_const_exprs` feature entirely, which has
wasted us many hours of productivity in the past. As much as it is a
cool feature in Rust it is still very much in an unstable state, leading
to a lot of `unconstrained generic const` errors, even though our arrays
were properly constrained.

This PR gets rid of `generic_const_exprs` in the following situations:
- within the `Stark` impl entirely. This ties closely with upstream
changes to the plonky2 evm - when reviewing, it would be useful to refer
to their changes too. It is pretty much the same, except we have an
extra public inputs var. See [this
PR](https://github.com/0xPolygonZero/plonky2/pull/1246) for their
changes. You will notice that I adopted the same stuff basically,
especially in our downstream fork of plonky2.
- within `HASH_SIZE` of the previously generic hasher, which was also
removed upstream previously,
[here](https://github.com/0xPolygonZero/plonky2/pull/1024).
- in other misc use cases, in `prep_table` for example.