suggestion(#751): test for program stark (#803)
See: https://github.com/0xmozak/mozak-vm/pull/751#discussion_r1375781714

There is a utility function called
`test_stark_circuit_constraints()`within `stark_testing` in `starky`
that helps ensure consistency between the native constraints and the
circuit constraints. We might want to add it here as well - it's not
that important in this current PR since the constraints are simple (only
a binary constraint), but it could serve as example code for future
circuit constraints that are more complex that we want to ensure behave
the same.