fix: ECALL should not change registers
In Risc-V ECALL does not have a destination register.  So for standard
Risc-V our decoder can not set one.  (Or at most could set a constant
default one for all ECALL.)

The way our constraints work is that a register can only change when it
is the destination register of an executed instruction.

But our IO-read ecall was returning the actual number of bytes
read in register 10.

For now, it's simplest to remove this feature, and let the VM assume
that the IO-read always fills the whole buffer it's given.