fix: ECALL should not change registers (#810)
In Risc-V ECALL does not have a destination register. So for standard
Risc-V our decoder can not set one. (Or at most could set a constant
default one for all ECALL.)

The way our constraints work is that a register can only change when it
is the destination register of an executed instruction.

But our IO-read ecall was returning the actual number of bytes read in
register 10.

For now, it's simplest to remove this feature, and let the VM assume
that the IO-read always fills the whole buffer it's given. (If we need
the feature later, we can think about how to re-enable it. Perhaps we
would return the information about the number of bytes written via
memory, eg by prefixing the length as `u32` to the buffer.)

Thanks and apologies to @supragya for finding this inconsistency during
his work on the protocol node.