feat(logup): introduce multiplicity view
Introduces `MultiplicityView` to both rangecheck and rangecheck limb -
this is a 'view' into the pair of columns `value` and `multiplicity`,
both of which should be present for logUp to work.

Note that this PR is opened as such in order to cater to the current
requirements of the trace to pass the tests, as well as introduce
additional columns (unconstrained intentionally for now) to support our
logUp efforts.