ci: add toolchain caching (#898)
This adds local caching to dramatically reduce the chances of [spurious
curl
failures](https://github.com/0xmozak/mozak-vm/actions/runs/6891290728/job/18746038046?pr=887)
(which seem to be some kind of throttling).

I tried to add remote caching of toolchains, but that turns out to be a
non-starter. Some of the binaries (e.g. `rustc`) depend on some
libraries which `rustup` also installs (to
`.rustup/toolchains/<toolchain>/lib`), but which libraries are installed
is environmentally dependent and differs between self-hosted and GitHub
hosted.

I created
[Daniel-Aaron-Bloom/local-cache](https://github.com/Daniel-Aaron-Bloom/local-cache)
and
[Daniel-Aaron-Bloom/rust-cache](https://github.com/Daniel-Aaron-Bloom/rust-cache)
as forks of other projects and improved them with high-quality, local,
shared caching support.

---------

Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>