refactor: switch `build.rs` generated/rust embedded elf binaries to simplify testing experience (#960)
We should automatically build the the elf binary and embed it instead of
requiring the caller to build it and load it at runtime. This change
adds a new helper crate to make all that simple.

This also means CI can remove a bunch of steps and easily test building
all the examples in one command.