XOR benches in perftool (#909)
Update: I decided to implement XOR and NOP benches separately for a
while. Will implement functionality to combine the plots in a separate
PR, and will have the plot comparing two benches featured there.

This PR introduces XOR and NOP benches. This is mainly to measure how
much extra overhead is incurred by a XOR instruction by comparing it to
same set of instructions with XOR replaced by NOP

---------

Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>