fix(merkleproof-trustedroot): write verification result to guest env
Ran into an issue with merkleproof-trustedroot going into an infinite
loop - doing a write with the boolean verification result (as a byte
array) fixes this for now.

Note that if the assertion fails then we'd run into an infinite loop and
that should now be interpreted as a failure to verify the merkle proof
in the example.

To test this PR, follow the README.md in the merkleproof-trustedroot
subdir.