CLI: make iotape inputs optional (#1034)
Unblocks #1002
This PR makes the iotape `Input` into `Option<Input>`, few
reasons:

1) With recent changes, the io tapes are now embedded within our
RISC-V binaries in the `.mozak_globals` section. We want to load from
them instead.
2) However! We may also want to provide the optionality of overriding
the embedded io tapes for debugging purposes (this would be the main use
case). This is when an io tape input into the cli is useful.
3) There are also certain programs/examples that require an input even
for the
native versions of the binaries.