Remove unused min_degree_bits (#1055)
The recursion threshold for a Plonky2 proof is 12. This means that when
we recursively verify a large proof, we obtain a large recursion proof.
We can continue to recursively verify this proof, but 2^12 is the
threshold beyond which we cannot further shrink the proof.

Polygon Zero uses the min_degree_bits parameter to determine when to end
the chain of proof recursion. We removed it because we are not using it
at the moment.