refactor(runner): use `Rc<[u8]>` instead of `Rc<Vec<u8>>` (#1113)
Extracted from #1098

There isn't a real point in this case in putting a `Vec` inside of an
`Rc`, since `Vec` itself is already a pointer and anything inside `Rc`
is immutable anyway. This also has the benefit of making the function
params explicit (we pass `Vec`s as params instead of doing a `to_vec` on
slices within initialization, which is opaque allocation done by the
caller).