feat: make tree recproof (#1114)
Adds two subcircuits, one for unbounded generic recursive proving, and a
one for proving construction of a merkle tree composed entirely of
multiple copies of the same public value.

The use of this later circuit isn't immediately apparent, but trust me,
it's useful as a part of a leaf circuit for an alternative state change
summarizing scheme. The [notion
page](https://www.notion.so/0xmozak/Proof-composition-and-rec-proofs-b3b64a80dc434b5d96561d59ed678d85)
will explain in more detail the specific usage.