refactor: remove callbacks from subcircuits (#1204)
follow up to #1188

This removes the callback structure in favor of a more typical approach.
This isn't needed for any of current circuits, but it will be needed for
some upcoming ones. This is largely moving code around, which does
result in the reordering of some calls, but should not otherwise change
functionality.