Make run_examples detect failure
`run_examples.sh` uses
```bash
cmd=$(.. some command that runs a test .. )
if [ ! "$cmd" ]; then
```
to try and detect whether the test ran successfully.

Alas, `cmd=$(.. command ..)` assigns the stdout of the command to the
valiable `cmd`.  It does not assign the exit code.

We could use `success=$?` here, but we don't actually need the extra
variable.