If outputs vec is empty() then halt with 0 (#1241)
`unwrap_unchecked()` in `env::finalize()` can result in `UNIMP` in
certain case. So this PR remove usage of them.

`UNIMP` is an alias for the `CSRRW` instruction:

> `UNIMP` : `C0001073`. This is an alias for `CSRRW x0, cycle, x0`.
Since `cycle` is a read-only CSR, then (whether this CSR exists or not)
an attempt to write into it will generate an illegal instruction
exception. This 32-bit form of `UNIMP` is emitted when targeting a
system without the C extension, or when the `.option norvc` directive is
used.