Make run_examples detect failure (#1240)
`run_examples.sh` uses
```bash
cmd=$(.. some command that runs a test .. )
if [ ! "$cmd" ]; then
```
to try and detect whether the test ran successfully.

Alas, `cmd=$(.. command ..)` assigns the stdout of the command to the
valiable `cmd`.  It does not assign the exit code.

We could use `success=$?` here, but we don't actually need the extra
variable.

Also: stop exiting successfully, just because we skipped any test.

Also: updated the command line argument passing in the script to be
compatible with changes in the CLI.

Also: skip `merkleproof-trustedroot` for now, because it's broken.