Revert "refactor: move util::execute to test_utils::execute (#1194)" (#1242)
This reverts commit caeb3412a91f43d09ca3c839d3d62dc510940334.

I need to revert it since:

1. Some parts of our codebase (currently wasm-demo) will use
`execute_code` function, so, `execute_code` will support `cfg(test)`
when it used in tests
2. Previously, before making all required changes to the
`starks/generation backend`, I thought that I can allow to
`execute_code` function to accept `ro/rw` addresses to overlap with
`mozak-ro-memory`. This was useful to allow unit-tests of LOAD & STORE
operation to randomly test all memory addresses. But, after realizing
that runner-execution code related to `mozak-ro-memory` needs to be
fixed by inserting of the `load` operations which lead to filtering
these execution rows in zero-init generation code, I realized that it
will no longer possible to accept memory overlapping at all. This is
because if we have overlapping address, this address will be filtered in
zero-init generation (since it belongs to the mozak-ro-memory), but
since we don't run real loader code, this address is not really used in
mozak-ro-memory --> execution rows wasn't extended --> CTL will fail
since in zero-init (Looking table) not such row, but in memory (Looked
table) there is such row. In read code execution, mozak-ro-memory init
stark will have this address, so CLT will not fail.

Bottom line - it was too early to do this refactoring of moving
`execute_code` to the tests....