feat: make default mozak ro memory mechanism (#1244)
1. Enable new io based on `mozak-ro-memory` 
2. Fixes memory-io tests that still use `ecall::IO` <-- this fix is
temporary - will be removed together with whole `ecall::IO`, the reason
we need to fix it is because these tests use io-tapes, which leads to
non-empty arguments, which leads to the situation when code-addresses
may lay inside `mozak-ro-memory` and this is forbidden. So, there is a
need to change `imm & offset` to be just `address` and randomize only
this single `address` variable. Also randomization filtering should be
used to make sure `address does not belong` to `mozak-ro-memory`

---------

Co-authored-by: bing <bingcicle@keemail.me>
Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>