fix: change some cli inputs to better handle `Option`s, fix run_examples script
This is a bit of a roundabout fix but we're skipping the token and
wallet programs first in the script, since they require a different way
to test them. `self_prog_id` and `system_tape` for example are required
inputs to the cli when running those programs.

It's out of scope to fix that in this PR so we'll skip them for now.