coretypes: define `args` and `ret` as struct `ArgsReturn`
Instead of just a tuple struct it is probably more appropriate to use a
normal struct. Reasoning: this struct has clearly defined `args` and
`ret` members, so we should prefer them instead of relying on `0` or
`1`. A tuple struct might be more appropriate for something like
coordinates (where it is known to be X and Y).