examples: wallet and token programs (#1282)
This PR includes the `wallet` and `token` programs used to demonstrate
the protocol PoC.

## For reviewers

### Summary

This is a big diff - but the below is a list of notable changes that you
should take note of as the reviewer:
- Inclusion of `self_prog_id`, `cast_list`, `call_tape` and `event_tape`
in linker script and various parts of the runner
- Updated memory trace generation to include the above new data.
- Added the `sdk` crate which has been a major WIP for a while now. The
`coretypes.rs` file is self-explanatory, but the `sys.rs` file contains
utilities and tools to interact with tapes for proving and verifying.
- Changed the `cli` to support `SystemTape` instead of separate private
and public tapes.
- And of course include the `wallet` and `token` programs.

### Wallet - native

The instructions to test the wallet program can be found in the README
of the project directory.

To understand the flow, look at `wallet/main_native.rs`:

All the wallet does is approve a signature (`ApproveSignature`) - and to
do that we do a `call_send()` which prepares the `call_tape` for proving
later. The native execution prepares the tapes for this reason. The
tapes are rkyv-serialized. (See [`rkyv`](https://github.com/rkyv/rkyv)
for more info on this serialization format)
 
### Token - native

The instructions to test the token program can be found in the README of
the project directory.

The token does a `Transfer` - which is also done wrapped in a
`call_send` that prepares the `call_tape`.

### MozakVM execution

From the above executions, we can prove/verify each mozakvm execution
separately with the tapes. The tapes come in the form of `SystemTape`,
which is rkyv-deserialized in a readable manner by the `sdk`.

Like `call_send()`, all the MozakVM execution does is do
`call_receive()`, which polls the `CallTape` for calls and checks that
execution went as intended. If at any point the calls do not match up,
proving will fail.

---------

Co-authored-by: Supragya Raj <supragyaraj@gmail.com>