sdk(prog_id): Use a single `Poseidon2HashType` to represent a `ProgramIdentifier`
This commit updates `Poseidon2HashType` to represent 32 bytes instead of
the current 4 bytes to be secure enough for production use; the previous
4 bytes were just for initial testing purposes.

This also entails changes with the `ProgramIdentifier`
type, which now is just represented by a single `Poseidon2HashType`, but
is still formed by the program rom hash, the memory init hash, and the
entrypoint like before - these values are hashed once to form a
program's `ProgramIdentifier`.

With this we also remove `hash.rs` from `guest` crate since we move its
functions over to the sdk now.