add: `native_helpers` for hint generation, set-up CI for SDK sanity (#1358)
Presents a simple function that only runs in native to generate the
index hints for `CanonicalEventTape`. In essence, since the
`CanonicalEventTape` is a sorted event tape according to address,
operation etc and may not be in the exact order as VM would have emitted
such events, `hints` in the form of indices of elements in sorted array
needs to be presented which `sort_with_hints` does.

Also, this PR presents CI elements to test SDK and keep it sane.