Use `i64` instead of field elements as co-efficients in the cross table lookup (WIP) (#1361)
Using `i64` is a lot more convenient, and we can convert them to field
elements (via `Field::from_noncanonical_i64`) at the end.

This drops a lot of now unnecessary type parameters.

I also switched to `checked_add` etc to implement the arithmetic
combination of `Column`s. The idea is that we don't want to be surprised
by any wrap-around for our coefficients, be that in `i64` or in field
elements. Especially for `reduce_with_powers` in
`circuits/src/linear_combination.rs` any wrap-around would be a security
vulnerability: that function is supposed to be generate linear
combinations that are injective over their domain. An unexpected
wrap-around would lead to clashes.

All the `checked_` arithmetic operations are restriction to
`circuits/src/linear_combination.rs`. Call-sites don't have to worry.