re-enable and test: Private and Public Input Tape(s) (#1403)
This PR re-introduces InputTapes and tests them via new SDK.

It also corrects wrong feature flag `rawio` and changes it to a better
name: `preinitmem_inputtape` which selects for pre-init memory based
input tapes. Not opting for this feature leads to use of `ecall` based
inputtapes, not fully implemented as of now.