circuits: deprecate flexbuffer serialization for serde_json
This is mainly to get rid of the dependencies that flexbuffer rely on.
But this also has the advantage of streamlining our serialization formats
as well.

We switch to `serde_json` for convenience since most our structs already
derive `serde`'s `Serialize` and `Deserialize` - we can consider
switching to `rkyv` in a future refactor if the need arises.