CTL: Compose linear transformations (#1450)
When we specify cross table lookups, we specify 'virtual' columns as
linear combinations of existing columns from our tables.

Afterwards, we generate challenges to fold a tuple of virtual columns
into a single virtual columns. That 'folding' is also a linear
combination.

This PR composes these two linear combinations, and then applies the
combination to our data, row by row.

That saves us a few field multiplications. My benchmarks show that it's
every so slightly faster on average, but I mostly want this as a
stepping stone for some other future optimisations I have in mind.