feat(circuits): introduce standalone mozak memory init tables
This commit introduces standalone tables for call tape, private tape,
public tape and event tape.

Previously, all these tables were put together in the same table, but
this makes it difficult to get the hash of each tape, which transaction
bundling requires (that requires the private tape hash and the call tape
hash).