cli: transaction bundling (#1335)
This PR adds the transaction bundling feature to the CLI which simulates
what a prover would do to prepare a transaction for verification by the
sequencer/verifier.

To make it easy to test, I included in commit
[6a27da9](https://github.com/0xmozak/mozak-vm/pull/1335/commits/6a27da9d2a543a6b767345b3f2fab8d2f4efa5d1)
a script to run/build the `wallet` and `token` bins in native/mozakvm
respectively. The script then runs the CLI with the feature implemented
in this PR to print the TX.

Currently, this PR prints the TX and dumps it in a json-serialized
string. You can test it using the script committed in this PR:

```sh
./scripts/test_bundle_tx.sh
```