feat: making sub tables of given stark table public (#1411)
The goal of the PR is to have a functionality to make a Subtable of
given stark table public.
This would be handy in future, when we have stark for individual tapes
(like event commitment tape) and want to make certain entries provably
visible to the verifier (Like the commitment hash).

To achieve this, the current approach we use is similar to that of CTL z
polynomial generation.
consider a subtable which we want to make public. The main idea is to
"compress" this subtable into a single field element, which verifier can
recreate on its own and compare it.
Given a subtable, we treat it exactly as an instance of "looking table"
in our CTL. That is, we "combine" its rows, and take the running sum of
inverses in its z polynomial. Now the compressed value would correspond
to opening of z polynomial in its last row.
 
The verifier can recreate this value, using the proposed subtable values
and ctl challenges. Then all it needs to do is check it against the z
polynomial's opening at the last row
 
# The design
`PublicSubTable` is used to mark the subtable which the prover wants to
make public
```rust
pub struct PublicSubTable {
    pub table: Table,
    pub num_rows: usize, // we need this, since it can't be obtained from `Table` alone
}
```
`MozakStark` marks all such subtables in `pub public_sub_tables:
&[PublicSubTable]` field, indicating all the sub tables that the prover
wants to make public.

## The task of prover
for each `PublicSubTable`, prover computes its corresponding `z`
polynomial in the same way as we do for a looking or looked table. That
is,
- `combine` each column by given `challenge`, if filter is present
- compute running sum of above values as `z` polynomial

Since the procedure is same as computation of `z` polynomial for CTL
tables, we decided to make this `z` polynomial part of `ctl_zs` polys,
and have it committed into `ctl_zs_commitment`. Further, since its
computation is preceded by commitment to trace polynomials, we can
safely reuse `ctl_challenges`. The information is stored in `CtlZData`
similarly.

Prover also computes the sub table's actual content,
`PublicSubTableValues`, and makes it part of `AllProof`
```rust
pub struct AllProof<F: RichField + Extendable<D>, C: GenericConfig<D, F = F>, const D: usize> {
    ...
    pub public_sub_table_values: TableKindArray<Vec<PublicSubTableValues<F>>>,  <---
}
```
## the task for verifier
The `z` polynomial correcteness is ensured via `CtlCheckVars`.
To ensure correctness of `PublicSubTableValues`, the verifier reduces it
using the `ctl_challenges` and checks it against the opening of
corresponding `z` polynomial at the last row.
```rust
for public_sub_table in public_sub_tables {
ensure!(
      reduced_public_sub_table_values_iter[public_sub_table.table.kind].next()
       == ctl_zs_openings[public_sub_table.table.kind].next()
);
}
```

## The Task of recursive verifier
Along with recursively encoding above computation, it makes
`PublicSubTableValues` public in its final proof. Hence The subtable has
been provably extracted out of its table.

```rust
for public_sub_table in &mozak_stark.public_sub_tables {
        builder.register_public_inputs(
            &public_sub_table_values_targets[public_sub_table.table.kind]
                .clone()
                .into_iter()
                .flatten()
                .flatten()
                .collect::<Vec<_>>(),
        );
    }
```

# Test 
(open to suggestions here)
Since at the moment, we don't have any sub table which we want to make
public,, it was hard to test it out. Hence I branched this PR and used a
dummy example (bitshift stark)
[here](https://github.com/0xmozak/mozak-vm/pull/1412) to test instead.

---------

Co-authored-by: Matthias Görgens <matthias.goergens@gmail.com>
Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>
Co-authored-by: bing <bingcicle@keemail.me>