Prepare CPU table to have holes in it later
Our CPU table is rather big.  If we want to move instruction out of it,
we need to prepare it for when it will have 'holes' in it, ie when the
clock doesn't just tick up from one row to the next.

Fortunately, almost all our CPU constraints already only deal with the
current row, `lv`.

In this PR we clean up and quarantine the few constraints that deal with
the next row, `nw`, into their own small section.

That introduces two new columns, but it will pay off later.  See
https://github.com/0xmozak/mozak-vm/pull/1461 for that.