cli(bundling): transaction bundling now only needs 1 plan
Due to some limitations, the previous CLI PR used 2 plans instead of 1,
which should be the correct approach.

Currently, we still only prove one program with this PR while waiting on
SDK extensions to support producing metadata of dependent programs
during the native execution of one program. This metadata will be used
for the CLI (and eventually the sequencer) to find where dependent ELFs
are compiled to be able to prove their execution(s) and be included in
the list of attestations.