feat+fix: redo Merge Subcircuit, add Merge Circuit (#1586)
Some of the the merge subcircuit interface didn't line up with how it
actually needs to be used in the VerifyProgram circuit:

- We need arbitrary presence, not just `hash == zero`
- We don't need `is_leaf` guarantees at all, manual control is
preferable (no recursion footguns)
- We really need a circuit, not a subcircuit

To the last point, organizationally we still want a subcircuit, but the
way we're going to use this, we really want to interface with a
completed `merge::BranchCircuit` and proof objects thereof. As such this
PR creates a new `recproofs/src/circuits/merge.rs` file and moves the
updated subcircuit into `recproofs/src/circuits/merge/core.rs`.