cli(bundling): transaction bundling now only needs 1 plan (#1584)
supercedes #1552

Due to some limitations, the previous CLI PR used 2 plans instead of 1,
which should be the correct approach.

Additionally, instead of a _plan_, we use the system tape as the plan
instead, because we have everything we need there.

Few changes you might notice:

- The first program is implied to be the entrypoint program. This is
where we get the call tape hash.
- We now only require `--system-tape <SYSTEM_TAPE>` to build a
transaction bundle. (You can test this with the test script committed in
this PR.)
- We only dump the tape and its debug version now. Removed `ProofBundle`
and the plan output completely.
- Important to note that we also rely entirely on the ECALL tables - not
the init tables for bundling.
- You may see some redundant usages of `RuntimeArguments`... these are
for preinit-memory, but since they are not of concern this PR, we will
just blindly use them for now (even if there's no effect), and remove
these usages in a followup PR.

---------

Co-authored-by: Matthias Görgens <matthias.goergens@mozak.com>