feat: Canonical hashing of EventTape in sdk (#1400)
This PR introduces a `Canonical` way to hash event tape, that is
ingested into recproofs. It has two components, which include hashing an
`Event` itself, and imposing an order on hashes and a way to `merkelize`
it to root hash.

### Canonical hashing an `Event`
This hash is obtained as flat hash  in that order, of
- `ProgramIdentifer` of event emitter which is `[u8; 32]`
- `EventType` represented as `[u8; 8]` (To be consistent with recproofs
for a while, namely `split_bytes` function, which is used to split a
`Target` into 8 bytes)
- `Address` of state object being referenced in the event. For now, to
be in sync with recproofs, we let this to be a `u64` represented as
`[u8; 8]` in little endian order.
- Hash of event's `value`, whichis `[u8; 32]`

The flat hash is without padding, since it fits into `WIDTH` already. We
don't worry about hash collisions, since we are more interested in
proving knowledge of preimage of hashes for this use case.

### Hashing the `Event` hashes themselves
For each `event`, their hashes are sorted according to following, in
that order
- `event.address` in increasing order.
- `event.type` and the following fields need not be ordered in any
certain way in order to be consistent with recproofs.
But to maintain uniformity, we impose the order given by enum. Note that
this is flexible as long as recproofs uses the same order.
```rust
pub enum EventType { Write, Ensure, Read, Create, Delete }
```
Once we have sorted the hashes, we proceed to `Merkleize` them. At each
level, we would have a node with address `addr` whose subtree contains
nodes which share their prefix to `addr`. Except for the nodes created
from the bottom layer, all other nodes would be created by hashing two
children nodes. At the bottom layer, we might have a group of hashes
which share the same `addr`, in which case, we create its parent node as
a merkle root of the group treated as leaves in that order.

---------

Co-authored-by: Matthias Görgens <matthias.goergens@gmail.com>
Co-authored-by: Supragya Raj <supragyaraj@gmail.com>